ログインユーザーと投稿者が一緒でない者が削除しようとする場合はインデックページへ

+++ app/controllers/prototypes_controller.rb
@@ -38,6 +38,7 @@ class PrototypesController < ApplicationController
   end
 
   def destroy
+    redirect_to action: :index unless current_user == @prototype.user
     prototype = Prototype.find(params[:id])
     prototype.destroy
     redirect_to root_path
